+++ common/pam.c
