+++ common/pjpeg-idct.c
