+++ common/pnm.c
