+++ common/pnm.h
